backend/requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
eventlet==0.35.0
pymongo==4.6.1
redis==5.0.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
pytest==7.4.3
